requirements.txt
dash
dash-bootstrap-components
pandas
pyarrow
scanpy
requests
Pillow
gunicorn